pyproject.toml
[tool.poetry]
name = "libsynomail"
version = "0.2.0"
description = ""
authors = ["Antonio Alonso <alloant@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
synochat = "^1.0.4"
synology-drive-api = "^1.0.14"
passlib = "^1.7.4"
openpyxl = "^3.1.2"
attrdict3 = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
